+++ src/components/index.ts
@@ -1,6 +1,7 @@
 export * from './ui'
 export * from './layout'
 export * from './mockups'
+export * from './mocksites'
 export * from './backgrounds'
 export * from './visuals'
 export * as ArtDirection from './art-direction'