+++ src/components/mocksites/index.ts
@@ -1,0 +1,2 @@
+export { CafeBadWebsite } from './CafeBadWebsite'
+export { CafeGoodWebsite } from './CafeGoodWebsite'